src/requirements.txt
PyMuPDF
pdfplumber
spacy
sentence-transformers
scikit-learn
pandas
numpy
Pillow
google-genai
tabulate
openai
python-dotenv 
vertexai
groq
rank-bm25
flask
pydantic